Fix bug in nested cons
Shout-out to @KCreate for noticing this. It turns out that

* `heap[CarOffset] = compile(car)`
* `heap[CdrOffset] = compile(cdr)`

does not make sense if `car` or `cdr` themselves modify `heap`. Instead
we need to store one of the intermediate results in a temporary value
(in this case `car`), compile `cdr` and store it into the heap, then fetch
`car` and store it into the heap.

Add tests.